Derivation of PDE in TEX format
I noticed that we didn't have a justification for why we are using this PDE to describe this motion, so I essentially transferred the presentation to TEX format. I tried to be brief since it's not super important but I might have went too low detail. A lot of this information would rely on some more physics/engineering concepts to derive though. This might not strictly be needed.