A partially formatted document for our submission
A lot of this writing is not my own, but copied from one of you all. I essentially saw that the document currently in here didn't have the abstract and such and so made that. Since the introduction written there seemed to fit the abstract more I put some of that in the abstract and then some in the introduction and tried to expand on the introduction, although that is not done. I also started work on the summary but I will need to see more of the results to finish that. This doesn't need to be used for any final submission but I just wanted to get down all this stuff somewhere.